requirements.txt
psycopg2<2.7.0
django<1.10
django-annoying<0.9.0
djangorestframework<3.4.0

##### Some optional programs, libs and such #####
# pillow==2.6.1
# pylibmc==1.3.0
# django-ckeditor==4.4.6
# django-docopt-command==0.2.0
django-admin-tools<0.8.0
django-pipeline<1.6.0
Jinja2<2.9

##### Testing with pytest #####
pytest-django==2.9.1
pytest-cov==2.2.0
factory-boy=2.6.0

# Disabled for now until south etc issues are investigated
# django-extensions<1.6.0

